Move reserved days to seperate state
