+++ client/src/components/CalendarMonth/CalendarMonth.jsx
@@ -44,14 +44,12 @@ CalendarMonth.propTypes = {
   reservedDays: PropTypes.arrayOf(PropTypes.string).isRequired,
 };
 
-const mapStateToProps = ({ availability, startingDay }, { month }) => {
-  const { reservedDays, minNights } = availability;
+const mapStateToProps = ({ reservedDays, startingDay }, { month }) => {
   const startingIndex = moment(month).diff(moment('2000-01-01'), 'days') - startingDay;
   const endingIndex = startingIndex + moment(month).daysInMonth();
   return {
     startingIndex,
     reservedDays: reservedDays.slice(startingIndex, endingIndex),
-    minNights,
   };
 };
