+++ client/src/store/reducer.js
@@ -1,11 +1,13 @@
 import { combineReducers } from 'redux';
 import availability from './reducers/availability.reducer';
+import reservedDays from './reducers/reservedDays.reducer';
 import checkInDay from './reducers/checkInDay.reducer';
 import currentDay from './reducers/currentDay.reducer';
 import startingDay from './reducers/startingDay.reducer';
 
 export default combineReducers({
   availability,
+  reservedDays,
   checkInDay,
   currentDay,
   startingDay,
